+++ anyblok/bloks/anyblok_core/system/sequence.py
@@ -1,6 +1,7 @@
 # This file is a part of the AnyBlok project
 #
 #    Copyright (C) 2015 Jean-Sebastien SUZANNE <jssuzanne@anybox.fr>
+#    Copyright (C) 2020 Jean-Sebastien SUZANNE <js.suzanne@gmail.com>
 #
 # This Source Code Form is subject to the terms of the Mozilla Public License,
 # v. 2.0. If a copy of the MPL was not distributed with this file,You can
@@ -68,7 +69,8 @@ class Sequence:
 
     id = Integer(primary_key=True)
     code = String(nullable=False)
-    number = Integer(nullable=False)
+    start = Integer(default=1)
+    current = Integer(default=None)
     seq_name = String(nullable=False)
     """Name of the sequence in the database.
 
@@ -127,7 +129,8 @@ class Sequence:
         :rtype: dict
         """
         seq_name = values.get('seq_name')
-        number = values.setdefault('number', 0)
+        start = values.setdefault('start', 1)
+
         if values.get("no_gap"):
             values.setdefault('seq_name', values.get("code", "no_gap"))
         else:
@@ -136,10 +139,7 @@ class Sequence:
                 seq_name = '%s_%d' % (cls.__tablename__, seq_id)
                 values['seq_name'] = seq_name
 
-            if number:
-                seq = SQLASequence(seq_name, number)
-            else:
-                seq = SQLASequence(seq_name)
+            seq = SQLASequence(seq_name, start=start)
             seq.create(cls.registry.bind)
         return values
 
@@ -161,10 +161,10 @@ class Sequence:
         """
         if self.no_gap:
             self.refresh(with_for_update={"nowait": True})
-            nextval = self.number + 1
+            nextval = self.current + 1 if self.current is not None else self.start
         else:
             nextval = self.registry.execute(SQLASequence(self.seq_name))
-        self.update(number=nextval)
+        self.update(current=nextval)
         return self.formater.format(code=self.code, seq=nextval, id=self.id)
 
     @classmethod